Collider Component added

+++ NULL Engine/Source/GameObject.cpp
@@ -159,7 +159,7 @@ bool GameObject::LoadState(ParsonNode& root)
 		if (!componentNode.NodeIsValid())
 			continue;
 		
-		ComponentType type	= (ComponentType)componentNode.GetNumber("Type");
+		ComponentType type	= (ComponentType)(int)componentNode.GetNumber("Type");
 
 		if (type == ComponentType::TRANSFORM)
 		{
